+++ CHANGES.rst
@@ -50,6 +50,9 @@ Release history
   ``Uniform`` but spreads points across the space evenly. (`#1611`_)
 - Added the ``ScatteredHypersphere`` distribution, which is similar to
   ``UniformHypersphere`` but spreads points across the space more evenly. (`#1611`_)
+- Added the ``RLS`` (recursive least-squares) learning rule, which is an online
+  version of the least-squares method typically used for offline decoder-solving.
+  (`#1611`_, `example <learn-product_>`__)
 
 **Changed**
 
@@ -122,6 +125,7 @@ Release history
 .. _#1629: https://github.com/nengo/nengo/pull/1629
 .. _#1630: https://github.com/nengo/nengo/pull/1630
 .. _#1640: https://github.com/nengo/nengo/pull/1640
+.. _learn-product: https://www.nengo.ai/nengo/examples/learning/learn-product.html
 
 3.0.0 (November 18, 2019)
 =========================